+++ RegistrationWizard.Domain/Repositories/ILocationRepository.cs
@@ -5,7 +5,7 @@ namespace RegistrationWizard.Domain.Repositories
     public interface ILocationRepository
     {
         Task<List<Country>> GetCountriesAsync(CancellationToken ct);
-        Task<List<Province>> GetCountryProvincesAsync(long countryId, CancellationToken ct);
+        Task<List<Province>> GetCountryProvincesAsync(int countryId, CancellationToken ct);
 
     }
 }
